+++ src/library/math/operator/hyper/exports.ts
@@ -1,1 +1,1 @@
-export * from './rank3';
+export * from './rank3/exports';
