+++ src/library/math/operator/hyper/rank3/exports.ts
